Reorganize processor factory utility classes
* Sources
* Sinks
* Processors
* DiagnosticProcessors
