+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/stream/impl/pipeline/PeekPipeline.java
@@ -18,7 +18,7 @@ package com.hazelcast.jet.stream.impl.pipeline;
 
 import com.hazelcast.core.IList;
 import com.hazelcast.jet.DAG;
-import com.hazelcast.jet.Processors;
+import com.hazelcast.jet.processor.Sinks;
 import com.hazelcast.jet.Vertex;
 import com.hazelcast.jet.stream.impl.StreamUtil;
 
@@ -42,7 +42,7 @@ class PeekPipeline<T> extends AbstractIntermediatePipeline<T, T> {
         String listName = uniqueListName();
         IList<T> list = context.getJetInstance().getList(listName);
         Vertex previous = upstream.buildDAG(dag);
-        Vertex writer = dag.newVertex("write-list-" + listName, Processors.writeList(listName));
+        Vertex writer = dag.newVertex("write-list-" + listName, Sinks.writeList(listName));
         if (upstream.isOrdered()) {
             writer.localParallelism(1);
         }
