+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/stream/impl/source/ListSourcePipeline.java
@@ -18,7 +18,7 @@ package com.hazelcast.jet.stream.impl.source;
 
 import com.hazelcast.core.IList;
 import com.hazelcast.jet.ProcessorMetaSupplier;
-import com.hazelcast.jet.Processors;
+import com.hazelcast.jet.processor.Sources;
 import com.hazelcast.jet.stream.impl.pipeline.AbstractSourcePipeline;
 import com.hazelcast.jet.stream.impl.pipeline.StreamContext;
 
@@ -33,7 +33,7 @@ public class ListSourcePipeline<E> extends AbstractSourcePipeline<E> {
 
     @Override
     protected ProcessorMetaSupplier getSourceMetaSupplier() {
-        return Processors.readList(list.getName());
+        return Sources.readList(list.getName());
     }
 
     @Override
