+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/EdgeTest.java
@@ -24,7 +24,7 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import static com.hazelcast.jet.Processors.noop;
+import static com.hazelcast.jet.processor.Processors.noop;
 import static com.hazelcast.jet.function.DistributedFunctions.wholeItem;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
