+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/impl/connector/HazelcastConnectorIntegrationTest.java
@@ -35,12 +35,12 @@ import org.junit.runner.RunWith;
 import java.util.concurrent.ExecutionException;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.Processors.readCache;
-import static com.hazelcast.jet.Processors.readList;
-import static com.hazelcast.jet.Processors.readMap;
-import static com.hazelcast.jet.Processors.writeCache;
-import static com.hazelcast.jet.Processors.writeList;
-import static com.hazelcast.jet.Processors.writeMap;
+import static com.hazelcast.jet.processor.Sources.readCache;
+import static com.hazelcast.jet.processor.Sources.readList;
+import static com.hazelcast.jet.processor.Sources.readMap;
+import static com.hazelcast.jet.processor.Sinks.writeCache;
+import static com.hazelcast.jet.processor.Sinks.writeList;
+import static com.hazelcast.jet.processor.Sinks.writeMap;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.IntStream.range;
 import static org.junit.Assert.assertEquals;
