+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/impl/connector/StreamTextSocketPTest.java
@@ -21,7 +21,7 @@ import com.hazelcast.jet.DAG;
 import com.hazelcast.jet.JetInstance;
 import com.hazelcast.jet.JetTestInstanceFactory;
 import com.hazelcast.jet.JetTestSupport;
-import com.hazelcast.jet.Processors;
+import com.hazelcast.jet.processor.Sources;
 import com.hazelcast.jet.Vertex;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
@@ -40,9 +40,9 @@ import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.Processors.noop;
-import static com.hazelcast.jet.Processors.streamTextSocket;
-import static com.hazelcast.jet.Processors.writeList;
+import static com.hazelcast.jet.processor.Processors.noop;
+import static com.hazelcast.jet.processor.Sources.streamTextSocket;
+import static com.hazelcast.jet.processor.Sinks.writeList;
 import static com.hazelcast.jet.impl.util.Util.uncheckRun;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -131,7 +131,7 @@ public class StreamTextSocketPTest extends JetTestSupport {
                 }
             })).start();
 
-            Vertex producer = new Vertex("producer", Processors.streamTextSocket(HOST, PORT)).localParallelism(1);
+            Vertex producer = new Vertex("producer", Sources.streamTextSocket(HOST, PORT)).localParallelism(1);
             Vertex sink = new Vertex("sink", noop()).localParallelism(1);
             DAG dag = new DAG()
                     .vertex(producer)
