+++ hazelcast-jet-hadoop/src/main/java/com/hazelcast/jet/connector/hadoop/ReadHdfsP.java
@@ -21,7 +21,7 @@ import com.hazelcast.jet.AbstractProcessor;
 import com.hazelcast.jet.Processor;
 import com.hazelcast.jet.ProcessorMetaSupplier;
 import com.hazelcast.jet.ProcessorSupplier;
-import com.hazelcast.jet.Processors;
+import com.hazelcast.jet.processor.Processors;
 import com.hazelcast.jet.Traverser;
 import com.hazelcast.jet.Util;
 import com.hazelcast.jet.function.DistributedBiFunction;
