+++ hazelcast-jet-hadoop/src/test/java/com/hazelcast/jet/connector/hadoop/ReadHdfsPTest.java
@@ -60,7 +60,7 @@ import java.util.concurrent.Future;
 import java.util.stream.IntStream;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.Processors.writeList;
+import static com.hazelcast.jet.processor.Sinks.writeList;
 import static com.hazelcast.jet.connector.hadoop.ReadHdfsP.readHdfs;
 import static com.hazelcast.jet.impl.util.Util.uncheckRun;
 import static java.util.stream.IntStream.range;
