+++ hazelcast-jet-kafka/src/test/java/com/hazelcast/jet/connector/kafka/WriteKafkaPTest.java
@@ -36,7 +36,7 @@ import java.util.concurrent.Future;
 import java.util.stream.Collectors;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.Processors.readMap;
+import static com.hazelcast.jet.processor.Sources.readMap;
 import static com.hazelcast.jet.connector.kafka.WriteKafkaP.writeKafka;
 import static java.util.stream.IntStream.range;
